tests for refuelings add_edit screen
