+++ lib/screens/add_edit/forms/barrel.dart
@@ -1,0 +1,1 @@
+export 'car_autocomplete.dart';
