+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   autocomplete_textfield:
     git:
       url: git://github.com/baylessj/flutter-autocomplete-textfield.git
+      ref: master
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
